Make "gp" compatible with new branches

+++ link/config/fish/functions/gp.fish
@@ -1,4 +1,8 @@
 function gp
   set BRANCH (git rev-parse --abbrev-ref HEAD)
-  git push --force-with-lease --set-upstream origin $BRANCH $argv
+  if git show-branch origin/$BRANCH > /dev/null 2>&1
+    git push --force-with-lease --set-upstream origin $BRANCH $argv
+  else
+    git push --set-upstream origin $BRANCH $argv
+  end
 end
